+++ JsonToolsNppPlugin/Tests/FormatPathTests.cs
@@ -9,7 +9,7 @@ namespace JSON_Tools.Tests
         public static bool Test()
         {
             JsonParser parser = new JsonParser();
-            JNode json = parser.Parse("{\"a\":1,\"b \":\"f\",\"cu\":[false,true,[{\"\":{\"bufrear\":null}}]],\"d'\":-1.5,\"b\\\\e\\\"\\r\\n\\t`\":[NaN,Infinity,-Infinity]}");
+            JNode json = parser.Parse("{\"a\":1,\"b \":\"f\",\"cu\":[false,true,[{\"\":{\"bufrear\":null}}]],\"d'\":-1.5,\"b\\\\e\\\"\\r\\n\\t`\":[NaN,Infinity,-Infinity], \"e\\\"'\": \"bar\"}");
             var testcases = new (int pos, KeyStyle style, string correctPath)[]
             {
                 (6, KeyStyle.JavaScript, ".a"),
@@ -36,6 +36,9 @@ namespace JSON_Tools.Tests
                 (93, KeyStyle.JavaScript, "['b\\\\e\"\\r\\n\\t`'][1]"),
                 (93, KeyStyle.RemesPath, "[`b\\\\e\"\\r\\n\\t\\``][1]"),
                 (93, KeyStyle.Python, "['b\\\\e\"\\r\\n\\t`'][1]"),
+                (121, KeyStyle.JavaScript, "[\"e\\\"'\"]"),
+                (121, KeyStyle.RemesPath,  "[`e\"'`]"),
+                (121, KeyStyle.Python,     "[\"e\\\"'\"]"),
             };
             int ii = 0;
             int testsFailed = 0;
@@ -59,18 +62,51 @@ namespace JSON_Tools.Tests
                     Npp.AddLine($"Got the path to position {pos} ({style} style) as {path}, but it should be {correctPath}");
                 }
             }
+
+            string separatorTestJsonStr = "{\"foo\": [12, {\"ba/r\": \"a\", \"baz\\\"\": null, \"12\": true}], \"baz\": {\"01\": false, \"a b\": null}, \"_f\": [-10,0.25]}";
+            JNode separatorTestJson = parser.Parse(separatorTestJsonStr);
+
+            var separatorTestcases = new (int pos, char separator, string correctPath)[]
+            {
+                (8, '/', "/foo"),
+                (8, ' ', " foo"),
+                (10, '/', "/foo/0"),
+                (23, '\\', "\\foo\\1\\\"ba/r\""),
+                (24, '\\', "\\foo\\1\\\"ba/r\""),
+                (39, '$', "$foo$1$\"baz\\\"\""),
+                (49, 'f', "f\"foo\"f1f\"12\""),
+                (75, 'z', "z\"baz\"z\"01\""),
+                (75, '/', "/baz/\"01\""),
+                (84, '/', "/baz/\"a b\""),
+                (88, ' ', " baz \"a b\""),
+                (100, '/', "/_f/0"),
+                (104, '_', "_\"_f\"_1"),
+            };
+            
+            foreach ((int pos, char separator, string correctPath) in separatorTestcases)
+            {
+                ii++;
+                string path;
+                try
+                {
+                    path = separatorTestJson.PathToPosition(pos, KeyStyle.JavaScript, separator);
+                }
+                catch (Exception ex)
+                {
+                    testsFailed++;
+                    Npp.AddLine($"While trying to get the path to position {pos} (Separator style, separator \"{separator}\"), threw exception\r\n{ex}");
+                    continue;
+                }
+                if (path != correctPath)
+                {
+                    testsFailed++;
+                    Npp.AddLine($"Got the path to position {pos} (Separator style, separator \"{separator}\") as {path}, but it should be {correctPath}");
+                }
+            }
+
             Npp.AddLine($"Failed {testsFailed} tests.");
             Npp.AddLine($"Passed {ii - testsFailed} tests.");
             return testsFailed > 0;
         }
-
-        //public void TestPathToTreeNode()
-        //{
-        //    TreeNode root = new TreeNode();
-        //    TreeView tree = new TreeView();
-        //    var pathsToJNodes = new Dictionary<string, JNode>();
-        //    TreeViewer.JsonTreePopulateHelper_DirectChildren(tree, root, json, pathsToJNodes);
-        //    var testcases = new ()
-        //}
     }
 }